% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-roxy.R
\docType{package}
\name{TSAOP-package}
\alias{TSAOP-package}
\alias{TSAOP}
\title{TSAOP: Time-Series Autoregressive Ordinal Probit}
\description{
Core infrastructure and native registration for the TSAOP package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mli171/TSAOP}
  \item Report bugs at \url{https://github.com/mli171/TSAOP/issues}
}

}
\author{
\strong{Maintainer}: Mo Li \email{mo.li@louisiana.edu}

}
